deps: add dependency on github.com/play-with-go/gitea
This PR simply adds the dependency on gitea; later PRs make changes to
the initial modules guide.

gitea is a module that encapsulates the infrastructure required to run a
private gitea instance that replicates repositories to a GitHub mirror
that sits beneath github.com/userguides/$repo. This infrastrcture will
sit behind the domain name play-with-go.dev, and be used as part of
guides that require remote VCS, e.g. when publishing modules.