cmd/controller: initial commit (#41)
The controller is a simple process that maps from a {guideName,
language, scenario} triple to the configuration for that setup. This
setup includes the presteps for the guide. When a guide is loaded, the
frontend first calls the controller with that triple (these details are
generated as part of the markdown output of preguide). The controller
then runs the presteps configured for that guide and returns the
additional environment the client should add in a call to PWD when
creating instances.