docker: fix site generate (#111)
The default jekyll image does some werid things when passed a JEKYLL_UID
and JEKYLL_GID value. It tries to modify the existing, build-created,
jekyll user in the container. However, this will not work if the passed
values happen to conflict with an existing user/group.

Instead, the entrypoint should define that JEKYLL_UID and JEKYLL_GID
represent the UID and GID as which the caller wants to run everything.
The jekyll user is really irrelevant in all of this.

Hence, we now provide a custom entrypoint and jekyll wrapper that
drives everything off JEKYLL_UID and JEKYLL_GID