deps: latest preguide to pick up extra checks for directives
This makes it much easier to detect issues in guides, but critically
ensures that the user will execute steps in the same order as the guide
script declares them. This is an important check, because we don't
(currently) run an independent check that the steps as referenced by the
prose result in a successful run.

This also brings in changes to improve the debug output of preguide,
which we now enable by default to make tracking down of some issues
easier.