deps: latest preguide to stop random output problems (#179)
Allocating a tty when we docker run a script ensures that stdout and
stderr are line buffered in the same way, and their output goes to the
same file descriptor. This should eradicate the random CI failures we
were seeing with output in random places (behaviour that could easily
have been seen locally too)

We also set TERM=dumb and NO_COLOR=true in order to prevent any escape
sequences from programs which think we want a richer experience because
the output is a tty.