guides: rewrite the "installing programs standalone" guide (#269)
This new feature, initially called "global install" when proposed, does
not have a formal name right now. "standalone install" is an okay name,
but it's pretty long and does not lend itself well to being an adverb,
like "install programs in a standalone way".

For now, use "directly", which seems to be pretty close to "while
ignoring go.mod". Rename the title and also the directory name.

As for the content itself, the guide is reorganized a bit to start with
the very basic "go get" example, moving to the "mktemp -d" variant, and
then finally to the 1.16 solution. This feels easier for the reader, who
is likely not aware of the history or why we arrived at "mktemp -d".

The other content changes are too minor to describe here, really.