guides: fix bad git add commands
This will automated at a later stage by play-with-go/preguide#153 (or
indeed by guides which are more verbose in their explanation to use,
e.g. running a git status after a git commit to show the state is
"clean") but for now add InformationOnly steps that verify we have a
clean working tree after creating a commit.

Fix a couple of instances where that was not the case.