docker: install a specific version of jekyll-toc
In our base image for the site part of Docker compose, we were running:

    RUN gem install jekyll-toc

This installed the latest versio of jekyll-toc at that point in time.
Which is a) clearly not reproducible, and b) liable to be brittle in the
presence of bad new versions.

Exactly that happened this morning with the release of v0.17.0, which
started to give rise to errors reported in:

    https://github.com/toshimaru/jekyll-toc/issues/141

This change installs a specific version of that dependency to make this
more resilient.